# server/openapi/auth.yaml
openapi: 3.1.0
info:
  title: Borneo Tracker Authentication API
  version: 1.0.0
  description: Same-origin opaque-session authentication. Non-GET browser requests require JSON, an approved Origin, and normally X-CSRF-Token. Logout without any auth/pre-auth cookie is an idempotent exception that clears cookies and returns 204.
servers:
  - url: http://localhost:3001
paths:
...
  /api/users/me:
    get:
      operationId: getMyProfile
      responses: { '200': { description: Current profile, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileResponse' } } } }, '401': { $ref: '#/components/responses/Error' } }
    patch:
      operationId: updateMyProfile
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileUpdate' } } } }
      responses: { '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileResponse' } } } }, '409': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
  /api/users/me/change-email:
    post:
      operationId: requestEmailChange
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [newEmail, currentPassword], properties: { newEmail: { $ref: '#/components/schemas/Email' }, currentPassword: { type: string, maxLength: 128 } } } } } }
      responses: { '202': { $ref: '#/components/responses/GenericAccepted' }, '401': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/RateLimited' } }
  /api/users/me/confirm-email-change:
    post:
      operationId: confirmEmailChange
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TokenRequest' } } } }
      responses: { '200': { description: Email changed and session rotated, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }, '400': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
components:
  parameters:
    Csrf: { name: X-CSRF-Token, in: header, required: true, schema: { type: string } }
  schemas:
    Email: { type: string, format: email, maxLength: 254 }
...
    TokenRequest: { type: object, additionalProperties: false, required: [token], properties: { token: { type: string, minLength: 40, maxLength: 100 } } }
...
    User: { type: object, additionalProperties: false, required: [id, email, firstName, lastName, role, emailVerified], properties: { id: { type: string, format: uuid }, email: { $ref: '#/components/schemas/Email' }, firstName: { type: string }, lastName: { type: string }, role: { enum: [USER, ADMIN] }, emailVerified: { type: boolean } } }
...
    LoginResponse: { type: object, required: [user, csrfToken], properties: { user: { $ref: '#/components/schemas/User' }, csrfToken: { type: string } } }
    ProfileUpdate: { type: object, additionalProperties: false, required: [version], properties: { version: { type: integer, minimum: 1 }, firstName: { type: string, minLength: 1, maxLength: 100 }, lastName: { type: string, minLength: 1, maxLength: 100 }, phoneCountryCode: { type: [string, 'null'], maxLength: 8 }, phoneNumber: { type: [string, 'null'], maxLength: 32 }, addressLine: { type: [string, 'null'], maxLength: 200 }, city: { type: [string, 'null'], maxLength: 100 }, state: { type: [string, 'null'], maxLength: 100 }, postalCode: { type: [string, 'null'], maxLength: 20 } } }
    ProfileResponse: { type: object, required: [user, profile, version], properties: { user: { $ref: '#/components/schemas/User' }, profile: { type: object }, version: { type: integer } } }
    Error: { type: object, required: [error], properties: { error: { type: object, required: [code, message, requestId], properties: { code: { type: string }, message: { type: string }, requestId: { type: string }, fieldErrors: { type: object, additionalProperties: { type: string } } } } } }
  responses:
    Error: { description: Stable API error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    GenericAccepted: { description: Generic response that does not disclose account existence, content: { application/json: { schema: { type: object, required: [message], properties: { message: { type: string } } } } } }
    RateLimited: { description: Rate limited, headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }